Use array as output from tectalic/openai package
Reason

As a misguided attempt (in 2023) to speed up php runtime, some PHP hosts are stripping comments from PHP files in production servers. Unfortunately, this configuration effectively broke any php code dependent on the Reflection API used on docblocks. It includes, amongst many other packages, the `symfony/routing`, `doctrine/dbal`, Laravel and so on.

The `tectalic/openai` package uses `spatie/data-transfer-object` to handle the casting (validation) of input and output of the library. Because it supports PHP 7.x versions, which have limited type support, the casting depends on docblock information.

Fortunately, there is a solution at the expense of catching some of the input/output errors: That library also works with plain old associated arrays.

Also: The side effect was that the PHPStan tool lost the knowledge of the structure of the results. So without defining it again, the solution is to relax the checking, hence introducing the `checkMissingIterableValueType: false` config.